requirements.txt
asgiref==3.8.1
Django==5.0.3
django-filter==24.2
djangorestframework==3.15.2
Faker==26.0.0
Markdown==3.6
python-dateutil==2.9.0.post0
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
validate-docbr==1.10.0
wheel==0.43.0
